On a Zen 4 processor:

Before:

```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean       505635 ns       505492 ns           10 Gigabytes=1.24931G/s docs=1.97828k/s
```

After:

```
------------------------------------------------------------------------------------------------
Benchmark                                      Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------
fast_minify_twitter/repeats:10_mean        29382 ns        29374 ns           10 Gigabytes=21.5023G/s docs=34.0488k/s
```


Fixes  https://github.com/simdjson/simdjson/issues/2334